Refactored valentineFormComponent, moved uploadUtils, stamp constants, and schema validation to other files
