+++ app/form/valentineFormComponent.tsx
@@ -14,19 +14,14 @@ import { Input } from "@/components/ui/input"
 import { Textarea } from "@/components/ui/textarea"
 import { Button } from "@/components/ui/button"
 import { useState } from "react"
-import { db, storage } from "@/lib/firebase"
-import { ref, uploadBytes, getDownloadURL } from "firebase/storage"
-import { collection, addDoc, updateDoc } from "firebase/firestore"
+import { db } from "@/lib/firebase"
+import { collection, addDoc } from "firebase/firestore"
 import { useRouter } from "next/navigation"
 import Link from "next/link"
 import Image from "next/image"
 import { MdHome } from "react-icons/md"
 import mofuFlower from "../../assets/mofu flower crop.png"
 import mofuHeart from "../../assets/mofu heart crop.png"
-import imageCompression from "browser-image-compression"
-import stamp1 from "../../assets/stamp 1.png"
-import stamp2 from "../../assets/stamp 2.png"
-import stamp3 from "../../assets/stamp 3.png"
 import stampFrame from "../../assets/square stamp frame.png"
 import placeholder from "../../assets/placeholder2.jpg"
 import ClickHeartEffect from "@/components/ClickHeartEffect"
@@ -34,46 +29,20 @@ import HeartBackground from "@/components/HeartBackground"
 import heic2any from "heic2any"
 import { logEvent } from "firebase/analytics"
 import { analytics } from "@/lib/firebase"
-
-const formSchema = z
-  .object({
-    senderName: z.string().min(1, "Your name is required"),
-    recipientName: z.string().min(1, "Valentine's name is required"),
-    message: z.string().min(1, "Message is required"),
-    image1: z.custom<File>().optional(),
-    caption1: z.string().optional(),
-    image2: z.custom<File>().optional(),
-    caption2: z.string().optional(),
-    selectedStamp: z.string().min(1, "Please select a stamp"),
-  })
-  .refine(
-    (data) => (data.image1 ? (data.caption1 ?? "").trim().length > 0 : true),
-    {
-      message: "You forgot to write a caption for your first image!",
-      path: ["caption1"],
-    },
-  )
-  .refine((data) => (data.caption1 ? !!data.image1 : true), {
-    message:
-      "You wrote a caption for your first image but forgot to upload the image!",
-    path: ["image1"],
-  })
-  .refine(
-    (data) => (data.image2 ? (data.caption2 ?? "").trim().length > 0 : true),
-    {
-      message: "You forgot to write a caption for your second image!",
-      path: ["caption2"],
-    },
-  )
-  .refine((data) => (data.caption2 ? !!data.image2 : true), {
-    message:
-      "You wrote a caption for your second image but forgot to upload the image!",
-    path: ["image2"],
-  })
+import { valentineFormSchema } from "@/schemas/valentineSchema"
+import { compressAndUploadImages } from "@/lib/uploadUtils"
+import { stamps } from "@/lib/constants"
 
 export default function ValentineForm() {
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
+  const [selectedStamp, setSelectedStamp] = useState<string | null>(null)
+  const [preview1, setPreview1] = useState<string | null>(null)
+  const [preview2, setPreview2] = useState<string | null>(null)
+  const [loading, setLoading] = useState(false)
+
+  const router = useRouter()
+
+  const form = useForm<z.infer<typeof valentineFormSchema>>({
+    resolver: zodResolver(valentineFormSchema),
     defaultValues: {
       senderName: "",
       recipientName: "",
@@ -84,52 +53,7 @@ export default function ValentineForm() {
     },
   })
 
-  const [selectedStamp, setSelectedStamp] = useState<string | null>(null)
-  const [preview1, setPreview1] = useState<string | null>(null)
-  const [preview2, setPreview2] = useState<string | null>(null)
-  const [loading, setLoading] = useState(false)
-
-  const stamps = [
-    { id: "stamp1", src: stamp1, alt: "Cats with cake" },
-    { id: "stamp2", src: stamp2, alt: "Two cats with heart" },
-    { id: "stamp3", src: stamp3, alt: "Penguin cats" },
-  ]
-
-  const router = useRouter()
-
-  async function uploadImage(file: File | undefined, path: string) {
-    if (!file) return null
-    const storageRef = ref(storage, path)
-    const snapshot = await uploadBytes(storageRef, file)
-    return await getDownloadURL(snapshot.ref)
-  }
-
-  interface CompressionOptions {
-    maxSizeMB: number
-    maxWidthOrHeight: number
-    useWebWorker: boolean
-  }
-
-  async function handleImageCompression(
-    imageFile: File,
-    compressionOptions: CompressionOptions,
-  ): Promise<File | null> {
-    try {
-      let fileToCompress: File = imageFile
-
-      // Compress the image
-      const compressedImage = await imageCompression(
-        fileToCompress,
-        compressionOptions,
-      )
-      return compressedImage
-    } catch (error) {
-      console.error("Image processing failed:", error)
-      return null
-    }
-  }
-
-  async function onSubmit(values: z.infer<typeof formSchema>) {
+  async function onSubmit(values: z.infer<typeof valentineFormSchema>) {
     setLoading(true)
 
     if (analytics) {
@@ -137,42 +61,18 @@ export default function ValentineForm() {
     }
 
     try {
-      let image1URL = null
-      let image2URL = null
-
       const compressionOptions = {
-        maxSizeMB: 0.5, // Target max size
-        maxWidthOrHeight: 800, // Max width/height
-        useWebWorker: true, // Improve performance
+        maxSizeMB: 0.5,
+        maxWidthOrHeight: 800,
+        useWebWorker: true,
       }
 
-      if (values.image1 && values.caption1) {
-        const compressedImage = await handleImageCompression(
-          values.image1,
-          compressionOptions,
-        )
-
-        if (compressedImage) {
-          image1URL = await uploadImage(
-            compressedImage,
-            `valentines/${values.senderName + Date.now()}/image1-${Date.now()}`,
-          )
-        }
-      }
-
-      if (values.image2 && values.caption2) {
-        const compressedImage = await handleImageCompression(
-          values.image2,
-          compressionOptions,
-        )
-
-        if (compressedImage) {
-          image2URL = await uploadImage(
-            compressedImage,
-            `valentines/${values.senderName + Date.now()}/image2-${Date.now()}`,
-          )
-        }
-      }
+      const [image1URL, image2URL] = await compressAndUploadImages(
+        [values.image1, values.image2],
+        "valentines",
+        values.senderName,
+        compressionOptions,
+      )
 
       const docRef = await addDoc(collection(db, "valentineMessages"), {
         senderName: values.senderName,
