+++ lib/constants.ts
@@ -1,3 +1,7 @@
+import stamp1 from "../assets/stamp 1.png"
+import stamp2 from "../assets/stamp 2.png"
+import stamp3 from "../assets/stamp 3.png"
+
 export const declineMessages = [
   "are you sure?",
   "really sure?",
@@ -14,4 +18,10 @@ export const declineMessages = [
   "wouldn't you reconsider?",
   "is that your final answer?",
   "you're breaking my heart ;(",
+]
+
+export const stamps = [
+  { id: "stamp1", src: stamp1, alt: "Cats with cake" },
+  { id: "stamp2", src: stamp2, alt: "Two cats with heart" },
+  { id: "stamp3", src: stamp3, alt: "Penguin cats" },
 ]
